Packaging refactor
