+++ src/main/java/com/mpkd/chatapp/domain/UserController.java
@@ -1,5 +1,6 @@
 package com.mpkd.chatapp.domain;
 
+import com.mpkd.chatapp.domain.dto.UserDTO;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
