+++ src/main/java/com/mpkd/chatapp/domain/UserFacade.java
@@ -1,5 +1,6 @@
 package com.mpkd.chatapp.domain;
 
+import com.mpkd.chatapp.domain.dto.UserDTO;
 import org.springframework.security.core.userdetails.UserDetailsService;
 
 public interface UserFacade extends UserDetailsService {
