+++ src/main/java/com/mpkd/chatapp/domain/dto/UserDTO.java
@@ -1,4 +1,4 @@
-package com.mpkd.chatapp.domain;
+package com.mpkd.chatapp.domain.dto;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
